+++ sys_service/sys_user.go
@@ -26,9 +26,9 @@ type (
 		GetSysUserById(ctx context.Context, userId int64) (*sys_entity.SysUser, error)
 		SetUserPermissionIds(ctx context.Context, userId int64, permissionIds []int64) (bool, error)
 		GetUserPermissionIds(ctx context.Context, userId int64) ([]int64, error)
-		SetUsername(ctx context.Context, newUsername string) (bool, error)
-		UpdateUserPassword(ctx context.Context, info sys_model.UpdateUserPassword) (bool, error)
-		ResetUserPassword(ctx context.Context, userId int64, password string, confirmPassword string) (bool, error)
+		SetUsername(ctx context.Context, newUsername string, userId int64) (bool, error)
+		UpdateUserPassword(ctx context.Context, info sys_model.UpdateUserPassword, loginUserName string) (bool, error)
+		ResetUserPassword(ctx context.Context, userId int64, password string, confirmPassword string, loginUser sys_entity.SysUser) (bool, error)
 	}
 )
